requirements.txt
torch
torchvision
numpy
scipy
scikit-image
matplotlib
jupyter
tifffile
zarr
numcodecs<0.16
Pillow
tqdm
tensorboard
albumentations
segmentation-models-pytorch
pytorch-lightning
timm
monai
wandb
pandas
opencv-contrib-python-headless
vesuvius
